+++ CHANGES.rst
@@ -15,7 +15,11 @@ Unreleased
   See the :ref:`feed-tags` user guide section for details.
   (:issue:`266`)
 * Allow running arbitrary actions before updating feeds
-  via :attr:`~Reader.before_feed_update_hooks`.
+  via :attr:`~Reader.before_feed_update_hooks`
+* The :mod:`~reader.plugins.mark_as_read` plugin now uses the
+  ``.reader.mark-as-read`` metadata for configuration.
+  Feeds using the old metadata, ``.reader.mark_as_read``,
+  will be migrated automatically on update until `reader` 3.0.
 * Add the :mod:`~reader._plugins.global_metadata` experimental plugin.